bot/requirements.txt
aiogram==3.18.0
python-dotenv==1.0.0
asyncpg==0.28.0
sqlalchemy==2.0.19
alembic==1.11.1
pydantic>=2.4.1,<2.11
loguru==0.7.0